+++ django_email_learning/personalised/views.py
@@ -672,6 +672,7 @@ class CertificateView(BaseTemplateView):
                     "courseTitle": certificate.enrollment.course.title,
                     "issueDate": certificate.issued_at.strftime("%B %d, %Y"),
                     "certificateNumber": certificate_number,
+                    "customFields": certificate.custom_fields or [],
                     "organizationName": certificate.enrollment.course.organization.name,
                     "logoUrl": certificate.enrollment.course.organization.logo.url
                     if certificate.enrollment.course.organization.logo
@@ -687,9 +688,6 @@ class CertificateView(BaseTemplateView):
                         ),
                         "issue_date": _("Issued on"),
                         "certificate_number": _("Certificate Number"),
-                        "organization_team": _("{organization_name} Team").format(
-                            organization_name=certificate.enrollment.course.organization.name
-                        ),
                     },
                 }
                 | self.get_app_context(),